Fix data type in api classes

+++ hazelcast/api/src/main/java/ar/edu/itba/pod/api/longestWait/LongestWaitMapper.java
@@ -4,9 +4,9 @@ import com.hazelcast.mapreduce.Context;
 import com.hazelcast.mapreduce.Mapper;
 
 //                                   LineaCSV                      pickUpLocationID
-public class LongestWaitMapper implements Mapper<Long, LongestWaitMapperValueIn, Long, LongestWaitReducerValue> {
+public class LongestWaitMapper implements Mapper<Long, LongestWaitMapperValueIn, Integer, LongestWaitReducerValue> {
     @Override
-    public void map(Long keyIn, LongestWaitMapperValueIn longestWaitMapperValueIn, Context<Long, LongestWaitReducerValue> context) {
+    public void map(Long keyIn, LongestWaitMapperValueIn longestWaitMapperValueIn, Context<Integer, LongestWaitReducerValue> context) {
         context.emit(longestWaitMapperValueIn.puLocationId(),
                 new LongestWaitReducerValue(longestWaitMapperValueIn.doLocationId(), longestWaitMapperValueIn.waitMillis(), longestWaitMapperValueIn.zoneName()));
     }
